requirements.txt
Django==4.2
django_cors_headers==3.14.0
djangorestframework==3.14.0
